Restore old similarity tests
