+++ tests/test_similarity_hamming.py
